+++ Core/Src/quadspi.c
@@ -224,7 +224,7 @@ HAL_StatusTypeDef PureCommand(uint16_t command)
     return HAL_OK;
 }
 
-HAL_StatusTypeDef Command_Rx_1DataLine_addr(uint16_t command, uint16_t *buffer, uint16_t addr, uint16_t size)
+HAL_StatusTypeDef Command_Rx_1DataLine_addr(uint16_t command, uint8_t *buffer, uint16_t addr, uint16_t size)
 {
     QSPI_CommandTypeDef sCommand = {0};
 
@@ -258,7 +258,7 @@ HAL_StatusTypeDef Command_Rx_1DataLine_addr(uint16_t command, uint16_t *buffer, 
     return HAL_OK;
 }
 
-HAL_StatusTypeDef Command_Rx_1DataLine(uint16_t command, uint16_t *buffer, uint16_t size, uint16_t dummyCycle)
+HAL_StatusTypeDef Command_Rx_1DataLine(uint16_t command, uint8_t *buffer, uint16_t size, uint16_t dummyCycle)
 {
     QSPI_CommandTypeDef sCommand = {0};
     sCommand.InstructionMode     = QSPI_INSTRUCTION_1_LINE;
@@ -291,7 +291,7 @@ HAL_StatusTypeDef Command_Rx_1DataLine(uint16_t command, uint16_t *buffer, uint1
     return HAL_OK;
 }
 
-HAL_StatusTypeDef Command_Rx_2DataLine(uint16_t command, uint16_t *buffer, uint16_t addr, uint16_t size)
+HAL_StatusTypeDef Command_Rx_2DataLine(uint16_t command, uint8_t *buffer, uint16_t addr, uint16_t size)
 {
     QSPI_CommandTypeDef sCommand = {0};
     sCommand.InstructionMode     = QSPI_INSTRUCTION_1_LINE;
@@ -324,7 +324,7 @@ HAL_StatusTypeDef Command_Rx_2DataLine(uint16_t command, uint16_t *buffer, uint1
     return HAL_OK;
 }
 
-HAL_StatusTypeDef Command_Tx_4DataLine(uint16_t command, uint16_t *buffer, uint16_t addr, uint16_t size)
+HAL_StatusTypeDef Command_Tx_4DataLine(uint16_t command, uint8_t *buffer, uint16_t addr, uint16_t size)
 {
     QSPI_CommandTypeDef sCommand = {0};
     sCommand.InstructionMode     = QSPI_INSTRUCTION_1_LINE;
@@ -358,7 +358,7 @@ HAL_StatusTypeDef Command_Tx_4DataLine(uint16_t command, uint16_t *buffer, uint1
     return HAL_OK;
 }
 
-HAL_StatusTypeDef StatusReg_Tx(uint16_t command, uint16_t regAddr, uint16_t data)
+HAL_StatusTypeDef StatusReg_Tx(uint16_t command, uint16_t regAddr, uint8_t data)
 {
     QSPI_CommandTypeDef sCommand = {0};
     sCommand.InstructionMode     = QSPI_INSTRUCTION_1_LINE;
@@ -387,7 +387,7 @@ HAL_StatusTypeDef StatusReg_Tx(uint16_t command, uint16_t regAddr, uint16_t data
     return HAL_OK;
 }
 
-HAL_StatusTypeDef StatusReg_Rx(uint16_t command, uint16_t regAddr, uint16_t *buffer)
+HAL_StatusTypeDef StatusReg_Rx(uint16_t command, uint16_t regAddr, uint8_t *buffer)
 {
     QSPI_CommandTypeDef sCommand = {0};
     sCommand.InstructionMode     = QSPI_INSTRUCTION_1_LINE;
